Allow basic ability to modify an inherited table
This update allows some basic position of table file entries
in the XML.  The new attribute "order='append'" is not possible
in the <dataTable>, <fieldTable>, and <diagTable> tags.  If one
of these tags has the "order" attribute, then the information in
the tag will be appended to the table.  This mainly allows the
inherited experiment's table to have new lines added.