Continue to create argFiles if unable to submit
If the ascii or restart output.stager job fails to submit (or run), then
the history argFile will never be created as the run script would exit
1.

This update changes this behavior by allowing the run script to continue
to the end if unable to submit more work.  A descriptive message is
printed in the stdout, and an email is sent to the user (if a batch job)
with instructions to get the job to continue.