Use store corrected opt_t in $hDate, and use hDate
Previous use of opt_t had a corrected opt_t stored back in as opt_t, and
then save opt_t as $hDate.  Then the two would be used interchangable
throughout the code.  This update leaves opt_t as the value entered via
the cli argument, and stores the corrected date in the hDate variable.

Issue #21.