Wrap Date::Manip::Date_Cmp
Created a wrapper routine FREUtil::date_Cmp for Date::Manip::Date_Cmp.

According to the Date::Manip documentation:

  [Date::Manip::Date_Cmp] currently does little more than use "cmp", but
  once the internal format for storing dates is in place (where time zone
  information is kept as part of the date), this routine will become more
  important.

As we do not deal with dates/times in different time zones, the need to
use Date::Manip::Date_Cmp over a simple cmp is not warranted (at this
time).  This routine is still put in place to allow for easier
modification to the routine if an issue is found.

Issue #21