Allow script to continue output.stager failed
These updates allowt he job run script to continue after the
output.stager job was unable to submit the save/transfer job.
This allows the restart/history ArgFile to be created if one of the
save/transfer output.stager calls fail to submit.  The output.retry
called during the next run will see the ArgFile and attempt to submit
the ArgFile.

Also, an error message is emailed to the user that states how to
maunally force the stager job to re-try the save/transfer.

Issue #19