Fix a sed command in csh section that used $
Special care needs to be taken when using '$' in a csh section in
frepp.  Several sed commands were added to remove the month and date
from dates, but the '$' in the search/replace regular expression
did not escape it, so Perl interpreted it.  This update fixes
the sed commands.

Issue #37