Update to not request ppn on pan
Pan does not need to request ppn when requesting nodes.  Moab on pan is
currently configured such that the number of available job slots does
not match the physical number of cores on the node, and the number of
slots may change as needed.

Pan is now also configured to use all 8 physical cores on a given node
when running parallel executables.