Update FRE.tool.modules.version.remote to match bronx version
The FRE.tool.modules.version.remote property set in
site/*/FRE.properties was set to 'test', which would load the fre/test
module on the remote site.  This recently caused end-to-end runs to fail
as fre-commands/test received updates that were not backwards
compatible, and caused some jobs to fail, as reported in issue #70.

This updates replaces test with the version of fre-commands used.
However, this update now needs to be applied on every fre-commands bronx
update.