#80 - Implement Curator triple IDs into frepp workflow
ISSUE:
 #80

PROBLEM:
 There is currently the desire to integrate tools with the Curator
database. As a result it is necessary to ingest experiments into the
database sooner than the first frepp -A call.

SOLUTION:
 This branch implements the call to fredb directly from frepp. This will allow the
experiment to be avaiable to the workflow with its triple IDs almost instantaneously.
The only drawback to this approach is a noticable delay of frepp when first getting
ingested into the database.

 This branch brings the following fre varaibles into existance:

   experID
   realizID
   runID
   tripleID

 There variables are aviable to both the analysis workflow and
the refineDiag workflow with the familiar 'set' syntax.