Update to use property 'MDBIswitch' for curator ingestion
PROBLEM:
The update to frepp meant that all experiments were put into the database regardless
of a use wanting to do so.

SOLUTION:
Implement getMDBIswitch sub routine to exptract the MDBIswitch from the XML and
make it a fre property

TESTING RESULTS:
Testing shows that an XML with MDBIswitch set to 'on' will be
ingested into the database, while others will not.