Add doF90Cpp attribute to compile tag
Add doF90Cpp attribute to the <compile> tag.  If this attribute is
added, and the value of the attribute is set to "yes", "true" or "on"
(case insensitive) then fremake will add in the extra csh code to
run cpp on the *.F90 files before attempting the fortran compiler.

Fixes #123