Capture git clone exit status
The original checkout.csh script did not capture the error status of the git clone.  This commit corrects that and will print the error messages with **GIT ERROR** prepended to each line.

This also has a fix for theia fre.properties file, and modifies the indentation of the file (an editor I used did that).

Fixes #138