Bronx-10 fix for time_bnds change in verona
This fix resolves the issue that causes some post-processing jobs to
time out due to a change in the field name from "time_bounds" to
"time_bnds".  A section of the frepp generated script checks to see if
the history file has a field with the name "*_bounds(".  If that field
doesn't exist, then the script attempts to add one.  The adding of this
field would cause some pp jobs to time out.

This fix resolves the attempt to add time_bounds if time_bounds or
time_bnds exists.

This also fixes a possible issue in seasonal pp jobs to use the correct
name of the variable (either time_bounds or time_bnds) as already
defined in the file.

Fixes #146.