Added allow_failure:true to gfdl-ws CI test
This is to allow the gfdl-ws tests to not affect the testing results.

This is safe (for now) as FRE is not ususally run on the workstations.

Once the Perl modules used by FRE on the workstations are updated (probably near when all workstations are on RHEL7), the allow_failure:true will be removed.