Remove theia from automated testing
ATT theia is not a supported platform, as we have been unable to keep up with system updates.  Although this may change in the future, I am configuring the gitlab CI to not run tests on theia.

As one can view, the theia job configurations are in the .gitlab-ci.yml file.  The job name now has a "." as the first character which essentially tell gitlab to not run those jobs.