Change to use project instead of projectInit in run script template
The projectInit shell variable is no longer used as it was originally.  The change caused several post-production jobs (e.g. output.stager and pp.starter) to be submitted to the wrong project on certain sites.  This modification renames projectInit => project, and removes other uses of projectInit from the runscript templates on other sites.  frerun was updated to no longer use projectInit, but instead project to set the project that should be used for the post-production jobs.